+++ Libs/pyqode/core/widgets/interactive.py
@@ -5,13 +5,14 @@ This module contains interactive widgets:
 """
 import locale
 import logging
-import os
 import sys
 
 from pyqode.core.api.client import PROCESS_ERROR_STRING
+from pyqode.core.managers.decorations import TextDecorationsManager
+from pyqode.core.managers.panels import PanelsManager
 from pyqode.qt.QtCore import Qt, Signal, QProcess, QProcessEnvironment
-from pyqode.qt.QtWidgets import QTextEdit
-from pyqode.qt.QtGui import QColor, QTextCursor, QFont
+from pyqode.qt.QtWidgets import QTextEdit, QAction, QApplication
+from pyqode.qt.QtGui import QColor, QTextCursor, QFont, QKeySequence
 
 
 def _logger():
@@ -35,12 +36,21 @@ class InteractiveConsole(QTextEdit):
           process started, process finished)
         - stderr_color: color of the process stderr
 
+    .. deprecated: Since v2.10.0, this widget is deprecated, you should use
+        :class:`pyqode.core.widgets.OutputWindow` instead.
+
     """
     #: Signal emitted when the process has finished.
     process_finished = Signal(int)
+    process_started = Signal()
 
     def __init__(self, parent=None):
         super(InteractiveConsole, self).__init__(parent)
+        self.panels = PanelsManager(self)
+        self.decorations = TextDecorationsManager(self)
+        from pyqode.core.panels import SearchAndReplacePanel
+        self.panels.append(SearchAndReplacePanel(),
+                           SearchAndReplacePanel.Position.TOP)
         self._stdout_col = QColor("#404040")
         self._app_msg_col = QColor("#4040FF")
         self._stdin_col = QColor("#22AA22")
@@ -63,6 +73,27 @@ class InteractiveConsole(QTextEdit):
         self._font_family = font
         self.setFont(QFont(font, 10))
         self.setReadOnly(True)
+        self._mask_user_input = False
+        action = QAction(_('Copy'), self)
+        action.setShortcut(QKeySequence.Copy)
+        action.triggered.connect(self.copy)
+        self.add_action(action)
+        action = QAction(_('Paste'), self)
+        action.setShortcut(QKeySequence.Paste)
+        action.triggered.connect(self.paste)
+        self.add_action(action)
+
+    def showEvent(self, event):
+        super(InteractiveConsole, self).showEvent(event)
+        self.panels.refresh()
+
+    def resizeEvent(self, e):
+        super(InteractiveConsole, self).resizeEvent(e)
+        self.panels.resize()
+
+    def add_action(self, action):
+        self.addAction(action)
+        action.setShortcutContext(Qt.WidgetShortcut)
 
     def set_writer(self, writer):
         """
@@ -82,15 +113,35 @@ class InteractiveConsole(QTextEdit):
             self._writer = writer
 
     def _on_stdout(self):
-        raw = self.process.readAllStandardOutput()
-        txt = bytes(raw).decode(locale.getpreferredencoding())
+        raw = bytes(self.process.readAllStandardOutput())
+        try:
+            txt = raw.decode(sys.getfilesystemencoding())
+        except UnicodeDecodeError:
+            txt = str(raw).replace("b'", '')[:-1].replace(
+                '\\r\\n', '\n').replace('\\\\', '\\')
+        _logger().debug('stdout: %s', txt)
         self._writer(self, txt, self.stdout_color)
 
     def _on_stderr(self):
-        txt = bytes(self.process.readAllStandardError()).decode(
-            locale.getpreferredencoding())
-        _logger().debug('%s', txt)
+        raw = bytes(self.process.readAllStandardError())
+        try:
+            txt = raw.decode(sys.getfilesystemencoding())
+        except UnicodeDecodeError:
+            txt = str(raw).replace("b'", '')[:-1].replace(
+                '\\r\\n', '\n').replace('\\\\', '\\')
+        _logger().debug('stderr: %s', txt)
         self._writer(self, txt, self.stderr_color)
+
+    @property
+    def exit_code(self):
+        if self.is_running:
+            return None
+        exit_status = self.process.exitStatus()
+        if exit_status == self.process.Crashed:
+            exit_code = 139
+        else:
+            exit_code = self.process.exitCode()
+        return exit_code
 
     @property
     def write_app_messages(self):
@@ -200,8 +251,21 @@ class InteractiveConsole(QTextEdit):
         """
         return self._running
 
+    @property
+    def mask_user_input(self):
+        return self._mask_user_input
+
+    @mask_user_input.setter
+    def mask_user_input(self, value):
+        """
+        If true, user input will be replaced by "*".
+
+        Could be useful to run commands as root.
+        """
+        self._mask_user_input = value
+
     def closeEvent(self, *args, **kwargs):
-        if self.process.state() == QProcess.Running:
+        if self.process and self.process.state() == QProcess.Running:
             self.process.terminate()
 
     def start_process(self, process, args=None, cwd=None, env=None):
@@ -227,6 +291,7 @@ class InteractiveConsole(QTextEdit):
         if not self._running:
             self.process = QProcess()
             self.process.finished.connect(self._on_process_finished)
+            self.process.started.connect(self.process_started.emit)
             self.process.error.connect(self._write_error)
             self.process.readyReadStandardError.connect(self._on_stderr)
             self.process.readyReadStandardOutput.connect(self._on_stdout)
@@ -256,52 +321,79 @@ class InteractiveConsole(QTextEdit):
         """
         Stop the process (by killing it).
         """
-        _logger().debug('killing process')
         if self.process is not None:
             self._user_stop = True
             self.process.kill()
             self.setReadOnly(True)
             self._running = False
 
+    def get_user_buffer_as_bytes(self):
+        """
+        Returns the user buffer as a bytes.
+        """
+        return bytes(self._usr_buffer, locale.getpreferredencoding())
+
     def keyPressEvent(self, event):
-        if event.key() == Qt.Key_Return or event.key() == Qt.Key_Enter:
+        ctrl = event.modifiers() & Qt.ControlModifier != 0
+        if not self.is_running or self.textCursor().hasSelection():
+            if event.key() == Qt.Key_C and ctrl:
+                self.copy()
+            return
+        propagate_to_parent = True
+        delete = event.key() in [Qt.Key_Backspace, Qt.Key_Delete]
+        if delete and not self._usr_buffer:
+            return
+        if event.key() == Qt.Key_V and ctrl:
+            # Paste to usr buffer
+            text = QApplication.clipboard().text()
+            self._usr_buffer += text
+            self.setTextColor(self._stdin_col)
+            if self._mask_user_input:
+                text = len(text) * '*'
+            self.insertPlainText(text)
+            return
+        if event.key() in [Qt.Key_Return, Qt.Key_Enter]:
             # send the user input to the child process
             if sys.platform == 'win32':
                 self._usr_buffer += "\r"
             self._usr_buffer += "\n"
-            self.process.write(bytes(
-                self._usr_buffer, locale.getpreferredencoding()))
+            self.process.write(self.get_user_buffer_as_bytes())
             self._usr_buffer = ""
         else:
-            if event.key() != Qt.Key_Backspace:
+            if not delete and len(event.text()):
                 txt = event.text()
                 self._usr_buffer += txt
+                if self._mask_user_input:
+                    txt = '*'
                 self.setTextColor(self._stdin_col)
-            elif event.text().isalnum():
-                self._usr_buffer = self._usr_buffer[
-                    0:len(self._usr_buffer) - 1]
+                self.insertPlainText(txt)
+                propagate_to_parent = False
+            elif delete:
+                self._usr_buffer = self._usr_buffer[:len(self._usr_buffer) - 1]
         # text is inserted here, the text color must be defined before this
         # line
-        super(InteractiveConsole, self).keyPressEvent(event)
+        if propagate_to_parent:
+            super(InteractiveConsole, self).keyPressEvent(event)
         self.setTextColor(self._stdout_col)
 
     def _on_process_finished(self, exit_code, exit_status):
+        if self is None:
+            return
+        self._running = False
         if not self._user_stop:
             if self._write_app_messages:
-                if exit_status == 0:
-                    # no crash, log exit code
-                    self._writer(
-                        self, "\nProcess finished with exit code %d" %
-                        exit_code, self._app_msg_col)
-                else:
-                    self._writer(
-                        self, "\nProcess finished by crashing "
-                              "(exit code not available)", self._app_msg_col)
-        self._running = False
+                self._writer(
+                    self, "\nProcess finished with exit code %d" %
+                    self.exit_code, self._app_msg_col)
         _logger().debug('process finished (exit_code=%r, exit_status=%r)',
                         exit_code, exit_status)
-        self.process_finished.emit(exit_code)
-        self.setReadOnly(True)
+        try:
+            self.process_finished.emit(exit_code)
+        except TypeError:
+            # pyqtSignal must be bound to a QObject, not 'InteractiveConsole'
+            pass
+        else:
+            self.setReadOnly(True)
 
     def _write_started(self):
         if not self._write_app_messages:
@@ -309,16 +401,17 @@ class InteractiveConsole(QTextEdit):
         self._writer(self, "{0} {1}\n".format(
             self._process_name, " ".join(self._args)), self._app_msg_col)
         self._running = True
-        _logger().debug('process started')
 
     def _write_error(self, error):
+        if self is None:
+            return
         if self._user_stop:
             self._writer(self, '\nProcess stopped by the user',
                          self.app_msg_color)
         else:
             err = PROCESS_ERROR_STRING[error]
             self._writer(self, "Error: %s" % err, self.stderr_color)
-            _logger().debug('process error: %s', err)
+            _logger().warn('process error: %s', err)
         self._running = False
 
     @staticmethod
@@ -336,10 +429,13 @@ class InteractiveConsole(QTextEdit):
         :param color: Desired text color
         :type color: QColor
         """
-        text_edit.moveCursor(QTextCursor.End)
-        text_edit.setTextColor(color)
-        text_edit.insertPlainText(text)
-        text_edit.moveCursor(QTextCursor.End)
+        try:
+            text_edit.moveCursor(QTextCursor.End)
+            text_edit.setTextColor(color)
+            text_edit.insertPlainText(text)
+            text_edit.moveCursor(QTextCursor.End)
+        except RuntimeError:
+            pass
 
     def apply_color_scheme(self, color_scheme):
         """
@@ -367,3 +463,12 @@ class InteractiveConsole(QTextEdit):
             self.stderr_color = QColor('#FF8080')
         else:
             self.stderr_color = QColor('red')
+
+
+if __name__ == '__main__':
+    from pyqode.qt import QtWidgets
+    app = QtWidgets.QApplication([])
+    console = InteractiveConsole()
+    console.start_process('cal')
+    console.show()
+    app.exec_()
